plug-ins: add Python 2/GIMP 2 API plug-ins not ported to new API.
After review by Elad Shahar, then by myself, we decided these 4 plug-ins
are probably not worth the effort to port them to the new GIMP 3 API
(and Python 3).
Either they are duplicate of existing script-fu script or of GEGL
operations or they are in unfinished state. All of them were not even
installed on stable builds.
See isssue gimp#4368.

Nevertheless as Akkana Peck (akk) points out on IRC, some of them are
interesting to study, in order to know how some things can be done in
GIMP plug-ins (because even with API changes, most of base code logics
is reusable) and also maybe these could be improved in the future to do
something actually useful or better than existing plug-ins. So I move
them to gimp-data-extras. I don't add them in the build system though,
because they probably should not be considered stable even in this
repository. But at least code is there for whoever wants to build good
plug-ins from there.