plug-ins: move py-slice.py from GIMP repository to gimp-data-extras.
Per polls on Twitter, Reddit, Patreon and Tipeee, most people were
willing to drop this plug-ins as not being relevant anymore (247 for
dropping and 63 for keeping it, so nearly 80% out of 310 participants).

So let's move it to gimp-data-extras allowing it to still have a
possible life, and who knows, maybe to even evolve.

While doing so, adding a --enable-gimp3 flag to the configure step. With
this option, the repository will look for a gimp-3.0 pkg-config file
instead of gimp-2.0 and will install to the relevant data and plug-in
folders.

Note that data (brushes, patterns, etc.) are installed both with
--enable-gimp3 and without, but the Python 3 plug-in is only installed
with the option enabled, since it is not GIMP 2.x compatible.