plug-ins: move benchmark-foreground-extract from GIMP repo.
We did port it to the new API but we are doubting it is useful at all.
It might have been, historically, when foreground extract algorithm was
first introduced, but probably not anymore, even more as the algorithm
is now implemented mostly on GEGL side anyway. So if any benchmark were
to be done, it should probably be added as GEGL tests.

Still we move it to the gimp-data-extras repository for being able to
easily bring it back in case we missed some use.