desktop: move the AppStream metadata from GIMP repository.
As far as I can see, this metadata file was added in 2017. See the
report: https://bugzilla.gnome.org/show_bug.cgi?id=763398
(renamed recently into a proper AppStream ID)

Unless I am missing something, it was created for the gimp-data-extras
package, never for anything shipped within the main GIMP repo. The
gimp-data-extras repo already existed back then, so I guess having
pushed it to the main repo was a mistake (?). Therefore moving it to its
rightful place so that this package gets proper metadata when listed in
application catalogs.

While doing this, also adding some initial localization. I know we are
being told to drop intltool and this could be a good start to do it as
the only localized data is this metadata, but I just wanted to make a
very quick data move, no thinking. If anyone wishes to make the port to
full-gettext (which is what various reports advises), we welcome merge
requests!