scripts: move Predator script from GIMP main repository.
The Predator script was broken because it relied on the plug-in-max-rgb, which
doesn't exist anymore in core GIMP. Rather than adding back the max-rgb plug-in,
it was ported and archived in gimp-data-extras (see previous commit), and so is
the Predator script. They are funny scripts, but we can't really say they are
absolute necessary base scripts for a raster editor.
Maybe later we can experiment with both these plug-ins to make them into
extensions and ship them again on the future extension platform.

In the main repo, this script was in plug-ins/script-fu/scripts/predator.scm

See report gimp#5440.